renamed controllers and view files
