+++ src/Controller/editDoctorController.java
@@ -1,4 +1,4 @@
-package sample;
+package Controller;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -8,7 +8,7 @@ import javafx.scene.control.TextField;
 
 import javax.swing.text.html.ImageView;
 
-public class Controller {
+public class editDoctorController {
     @FXML
     Button logout;
     @FXML
