+++ src/View/editRoomView.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.paint.*?>
 <?import javafx.scene.text.*?>
 
-<AnchorPane id="AnchorPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/2.2">
+<AnchorPane id="AnchorPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/2.2" fx:controller="Controller.editRoomController">
     <children>
         <Button id="logout" fx:id="logoutButton" layoutX="14.0" layoutY="20.0" mnemonicParsing="false" text="Logout" />
         <Button fx:id="backButton" layoutX="14.0" layoutY="366.0" mnemonicParsing="false" text="Back" />
